# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BlindCash API
  version: 1.0.0
  description: |
    BlindCash — 基于 Chaum 盲签名的 eCash 系统 API。

    **协议概述**（M2-M7）：
    - 取款走 4-move cut-and-choose 盲签名协议：init → submit → reveal → cancel
    - 支付走 token 兑付：POST /api/payment
    - 银行公钥可公开获取：GET /api/bank/pubkey

    **认证**：所有受保护接口需 `Authorization: Bearer <JWT>`。
    JWT 由 /api/auth/register 或 /api/auth/login 返回。

    **限流**（Phase 4）：
    - 全局：100 req/min/IP
    - 登录/注册：5 req/min/IP（防爆破）
    - 取款/支付：10 req/min/user

    **文献参考**：
    - [1] D. Chaum, "Blind Signatures for Untraceable Payments", CRYPTO 1982
    - [2] BIP-340 — Schnorr 签名（secp256k1）
    - [3] NIST SP 800-57rev5 §8.3.4 — 密钥轮换 cryptoperiod
servers:
  - url: http://localhost:4100
    description: 本地开发
  - url: /
    description: 相对路径（生产同源）
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误码
        message:
          type: string
        req_id:
          type: string
          description: 请求 ID（Phase 4 错误响应带此值便于追溯）
security: []
paths:
...
  /api/bank/reserve:
    get:
      summary: 查银行准备金状态（admin）
      security:
        - bearerAuth: []
      description: 管理员查看货币发行/回收/准备金状态。需 admin 角色。
      responses:
        '200':
          description: 准备金数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_issued:
                    type: integer
                  total_redeemed:
                    type: integer
                  reserve_balance:
                    type: integer
                  in_flight:
                    type: integer
                    description: 在途取款金额
                  sum_balance:
                    type: integer
                  updated_at:
                    type: string
        '401':
          description: 未认证
        '403':
          description: 非 admin
  /api/bank/deposit:
    post:
      summary: 充值（simulated fiat rail）
      security:
        - bearerAuth: []
      description: |
        自助充值：模拟外部法币入账。
        单次上限 1000，24h 滚动累计上限 5000。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: integer
                  minimum: 1
                  maximum: 1000
      responses:
        '200':
          description: 充值成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  deposited:
                    type: integer
                  new_balance:
                    type: integer
        '400':
          description: 金额无效
  /api/bank/redeem:
    post:
      summary: 退币（兑付 token 到自己账户）
      security:
        - bearerAuth: []
      description: |
        用户将取款得到的 token 兑付到自己账户。
        与 /api/payment 共享 spent_coins 表，同一 token 只能消费一次。
        key_id 透传给 service 层用 getPublicKeyByVersion 查对应公钥验签。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serial, amount, R_prime, s_prime]
              properties:
                serial:
                  type: string
                amount:
                  type: integer
                R_prime:
                  type: string
                s_prime:
                  type: string
                key_id:
                  type: integer
      responses:
        '200':
          description: 兑付成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  deposited:
                    type: integer
                  new_balance:
                    type: integer
        '400':
          description: token 格式错误或签名无效
        '409':
          description: token 已被消费（双花）
  /api/bank/redeem-split:
    post:
      summary: 找零兑付（Phase 6.2）
      security:
        - bearerAuth: []
      description: |
        将大面额 token 全额退到账户，并在 spent_coins 表按 split_denomination
        记账（教学化展示"等价于多少枚小币"）。用户随后可另走 4-move 取款流程
        取 split_denomination 面额的新 token。

        **隐私局限**（返回 limitations 数组，前端必须诚实展示）：
        - 不是真正的 Chaum 找零（银行只全额退币，需另走取款流程）
        - spent_coins 按 split_denomination 记账仅是教学展示
        - 时间侧信道：大额兑付 + 后续小额取款可被关联（参考 Chaum 1985）

        split_denomination 必须能整除 amount。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serial, amount, R_prime, s_prime, split_denomination]
              properties:
                serial:
                  type: string
                amount:
                  type: integer
                R_prime:
                  type: string
                s_prime:
                  type: string
                key_id:
                  type: integer
                  description: 银行密钥版本号（token v2，可选）
                split_denomination:
                  type: integer
                  enum: [1, 5, 10, 50, 100]
                  description: 拆分目标面额，必须能整除 amount
      responses:
        '200':
          description: 找零兑付成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  deposited:
                    type: integer
                  new_balance:
                    type: integer
                  split_denomination:
                    type: integer
                  split_count:
                    type: integer
                    description: amount / split_denomination
                  limitations:
                    type: array
                    items:
                      type: string
                    description: 隐私局限说明，前端必须展示
        '400':
          description: |
            错误码：
            - INVALID_SPLIT_DENOMINATION — split_denomination 不在 [1,5,10,50,100]
            - SPLIT_NOT_DIVISIBLE — amount 不能被 split_denomination 整除
            - SIGNATURE_INVALID — 签名验证失败
            - MALFORMED_TOKEN — token 格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: token 已被消费（双花）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/audit:
    get:
      summary: 查审计日志（admin）
      security:
        - bearerAuth: []
      description: 管理员分页查询 audit_log 表。需 admin 角色。
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
        - in: query
          name: action
          schema:
            type: string
          description: 按 action 过滤（如 deposit, withdraw, key_rotate）
      responses:
        '200':
          description: 审计日志
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
        '403':
          description: 非 admin
...
  /api/admin/rotate-key:
    post:
      summary: 密钥轮换（admin）
      security:
        - bearerAuth: []
      description: |
        旧密钥标 retired + retired_until=now+90d，生成新 active 密钥。
        需 admin 角色。
      responses:
        '200':
          description: 轮换成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  old_version:
                    type: integer
                  new_version:
                    type: integer
        '403':
          description: 非 admin